+++ src/utils.js
@@ -1,5 +1,8 @@
 /* eslint camelcase: ["off"] */
 'use strict';
+
+const castArray = require('lodash.castarray');
+
 /**
  * Represents an error result received from a TP-Link device.
  *
@@ -85,11 +88,16 @@ function normalizeMac (mac = '') {
   return mac.replace(/[^A-Za-z0-9]/g, '').toUpperCase();
 }
 
-function compareMac (mac = '', macPattern = '') {
-  let re = new RegExp(
-    macPattern.replace(/[^A-Za-z0-9*]/g, '').replace(/[*]/g, '.').toUpperCase()
-  );
-  return re.test(normalizeMac(mac));
+function compareMac (mac = '', macPattern) {
+  const macPatterns = castArray(macPattern).map((p) => {
+    return new RegExp('^' +
+      p.replace(/[^A-Za-z0-9?*]/g, '')
+        .replace(/[?]/g, '.')
+        .replace(/[*]/g, '.*')
+        .toUpperCase() + '$');
+  });
+  const normalizedMac = normalizeMac(mac);
+  return (macPatterns.findIndex((p) => p.test(normalizedMac)) !== -1);
 }
 
 module.exports = {